STY: Moved create_files to test_files
